cleanup

+++ src/categories.cpp
@@ -188,7 +188,7 @@ int CategoryFactory::addCategory(const QString &name, const std::vector<int> &ne
 
 void CategoryFactory::addCategory(int id, const QString &name, const std::vector<int> &nexusIDs, int parentID)
 {
-  int index = m_Categories.size();
+  int index = static_cast<int>(m_Categories.size());
   m_Categories.push_back(Category(index, id, name, nexusIDs, parentID));
   for (int nexusID : nexusIDs) {
     m_NexusMap[nexusID] = index;
