+++ src/logbuffer.cpp
@@ -53,13 +53,15 @@ void LogBuffer::logMessage(QtMsgType type, const QString &message)
   if (type >= m_MinMsgType) {
     Message msg = {type, QTime::currentTime(), message};
     if (m_NumMessages < m_Messages.size()) {
-      beginInsertRows(QModelIndex(), m_NumMessages, m_NumMessages + 1);
+      beginInsertRows(QModelIndex(), static_cast<int>(m_NumMessages),
+                      static_cast<int>(m_NumMessages) + 1);
     }
     m_Messages.at(m_NumMessages % m_Messages.size()) = msg;
     if (m_NumMessages < m_Messages.size()) {
       endInsertRows();
     } else {
-      emit dataChanged(createIndex(0, 0), createIndex(m_Messages.size(), 0));
+      emit dataChanged(createIndex(0, 0),
+                       createIndex(static_cast<int>(m_Messages.size()), 0));
     }
     ++m_NumMessages;
     if (type >= QtCriticalMsg) {
@@ -84,9 +86,10 @@ void LogBuffer::write() const
     return;
   }
 
-  unsigned int i = (m_NumMessages > m_Messages.size())
-                       ? m_NumMessages - m_Messages.size()
-                       : 0U;
+  unsigned int i
+      = (m_NumMessages > m_Messages.size())
+            ? static_cast<unsigned int>(m_NumMessages - m_Messages.size())
+            : 0U;
   for (; i < m_NumMessages; ++i) {
     file.write(m_Messages.at(i % m_Messages.size()).toString().toUtf8());
     file.write("\r\n");
@@ -165,7 +168,7 @@ int LogBuffer::rowCount(const QModelIndex &parent) const
   if (parent.isValid())
     return 0;
   else
-    return std::min(m_NumMessages, m_Messages.size());
+    return static_cast<int>(std::min(m_NumMessages, m_Messages.size()));
 }
 
 int LogBuffer::columnCount(const QModelIndex &) const
@@ -175,9 +178,10 @@ int LogBuffer::columnCount(const QModelIndex &) const
 
 QVariant LogBuffer::data(const QModelIndex &index, int role) const
 {
-  unsigned offset = m_NumMessages < m_Messages.size()
-                        ? 0
-                        : m_NumMessages - m_Messages.size();
+  unsigned int offset
+      = m_NumMessages < m_Messages.size()
+            ? 0
+            : static_cast<unsigned int>(m_NumMessages - m_Messages.size());
   unsigned int msgIndex = (offset + index.row() + 1) % m_Messages.size();
   switch (role) {
     case Qt::DisplayRole: {
